core: use safeclib in server_object_create
The code changes in this PR were automatically generated by the
Permanence AI Coder and reviewed by @jweese, @fwph and @eelenberg. This
changes the `strcpy` call to use `strcpy_s` provided by the SafeC
library. Here `(*obj)->name` is an array of known size so the use of
`sizeof` is valid.

on-behalf-of: @permanence-ai <github-ai@permanence.ai>